COMPAT: Remove warnings caused by pandas 0.17.0
